ewijoi

+++ src/components/buffer-clients/buffer-client.service.ts
@@ -105,31 +105,35 @@ export class BufferClientService {
     }
 
     async joinchannelForBufferClients(): Promise<string> {
-        console.log("Joining Channel Started")
-        await this.telegramService.disconnectAll();
-        await sleep(2000);
-        const clients = await this.bufferClientModel.find({ channels: { "$lt": 180 } }).limit(4);
-        clients.map(async (document) => {
-            try {
-                const client = await this.telegramService.createClient(document.mobile, false, false);
-                console.log("Started Joining for : ", document.mobile)
-                const channels = await client.channelInfo(true);
-                console.log("Existing Channels Length : ", channels.ids.length);
-                await this.update(document.mobile, { channels: channels.ids.length });
-                const keys = ['wife', 'adult', 'lanj', 'lesb', 'paid', 'coupl', 'cpl', 'randi', 'bhab', 'boy', 'girl', 'friend', 'frnd', 'boob', 'pussy', 'dating', 'swap', 'gay', 'sex', 'bitch', 'love', 'video', 'service', 'real', 'call', 'desi'];
-                const result = await this.activeChannelsService.getActiveChannels(150, 0, keys, channels.ids);
-                this.joinChannelMap.set(document.mobile, result);
-                await this.telegramService.deleteClient(document.mobile);
-                // console.log("DbChannelsLen: ", result.length);
-                // let resp = '';
-                // this.telegramService.joinChannels(document.mobile, result);
-            } catch (error) {
-                parseError(error)
+        if (!this.telegramService.getActiveClientSetup()) {
+            console.log("Joining Channel Started")
+            await this.telegramService.disconnectAll();
+            await sleep(2000);
+            const clients = await this.bufferClientModel.find({ channels: { "$lt": 180 } }).limit(4);
+            for (const document of clients) {
+                try {
+                    const client = await this.telegramService.createClient(document.mobile, false, false);
+                    console.log("Started Joining for : ", document.mobile)
+                    const channels = await client.channelInfo(true);
+                    console.log("Existing Channels Length : ", channels.ids.length);
+                    await this.update(document.mobile, { channels: channels.ids.length });
+                    const keys = ['wife', 'adult', 'lanj', 'lesb', 'paid', 'coupl', 'cpl', 'randi', 'bhab', 'boy', 'girl', 'friend', 'frnd', 'boob', 'pussy', 'dating', 'swap', 'gay', 'sex', 'bitch', 'love', 'video', 'service', 'real', 'call', 'desi'];
+                    const result = await this.activeChannelsService.getActiveChannels(150, 0, keys, channels.ids);
+                    this.joinChannelMap.set(document.mobile, result);
+                    await this.telegramService.deleteClient(document.mobile);
+                    // console.log("DbChannelsLen: ", result.length);
+                    // let resp = '';
+                    // this.telegramService.joinChannels(document.mobile, result);
+                } catch (error) {
+                    parseError(error)
+                }
             }
-        })
-        this.joinChannelQueue();
-        console.log("Joining Channel Triggered Succesfully for ", clients.length);
-        return "Initiated Joining channels"
+            this.joinChannelQueue();
+            console.log("Joining Channel Triggered Succesfully for ", clients.length);
+            return "Initiated Joining channels"
+        } else {
+            console.log("ignored active check buffer channels as active client setup exists")
+        }
     }
 
     async joinChannelQueue() {
@@ -137,8 +141,7 @@ export class BufferClientService {
             const keys = Array.from(this.joinChannelMap.keys());
             if (keys.length > 0) {
                 console.log("In JOIN CHANNEL interval: ", new Date().toISOString());
-
-                const promises = keys.map(async mobile => {
+                for (const mobile of keys) {
                     const channels = this.joinChannelMap.get(mobile);
                     if (channels && channels.length > 0) {
                         const channel = channels.shift();
@@ -155,8 +158,7 @@ export class BufferClientService {
                     } else {
                         this.joinChannelMap.delete(mobile);
                     }
-                });
-                await Promise.all(promises);
+                }
             } else {
                 this.clearJoinChannelInterval()
             }
